Worked on load/save routines.
The new icd format seems to work.

+++ src/model/tools/fill_imp.rs
@@ -160,7 +160,7 @@ impl Tool for FillTool {
 
     fn handle_click(&mut self, editor: &mut AnsiEditor, button: i32, pos: Position) -> Event {
         if button == 1 {
-            if editor.cur_layer >= editor.buffer_view.lock().buf.layers.len() as i32 {
+            if editor.cur_layer >= editor.buffer_view.lock().buf.layers.len() {
                 return Event::None;
             }
             let attr = editor.buffer_view.lock().caret.get_attribute();
